services/txt-passthrough/pyproject.toml
[project]
name = "txt-passthrough"
version = "0.1.0"
description = "TXT passthrough service - moves TXT files from parseable to processed bucket"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aio-pika>=9.5.5",
    "minio>=7.2.15",
    "python-dotenv>=1.1.0",
]
